changed directories and names
